sdcc_recommendation_service/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
cachetools==4.1.1         # via google-auth
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
contextvars==2.4          # via opencensus-context
google-api-core==1.21.0   # via -r requirements.in, google-api-python-client, opencensus
google-api-python-client==1.10.0  # via google-cloud-profiler, google-python-cloud-debugger
google-auth-httplib2==0.0.4  # via google-api-python-client, google-cloud-profiler, google-python-cloud-debugger
google-auth==1.19.2       # via google-api-core, google-api-python-client, google-auth-httplib2, google-cloud-profiler, google-python-cloud-debugger
google-cloud-profiler==2.0.0  # via -r requirements.in
google-python-cloud-debugger==2.15  # via -r requirements.in
googleapis-common-protos==1.52.0  # via google-api-core
grpcio==1.30.0            # via -r requirements.in
httplib2==0.18.1          # via google-api-python-client, google-auth-httplib2
idna==2.10                # via requests
immutables==0.14          # via contextvars
joblib==0.16.0            # via scikit-learn
numpy==1.19.0             # via -r requirements.in, pandas, scikit-learn, scipy
opencensus-context==0.1.1  # via opencensus
opencensus[stackdriver]==0.7.10  # via -r requirements.in
pandas==1.0.5             # via -r requirements.in
protobuf==3.12.2          # via google-api-core, google-cloud-profiler, googleapis-common-protos
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
python-dateutil==2.8.1    # via pandas
python-json-logger==0.1.11  # via -r requirements.in
pytz==2020.1              # via google-api-core, pandas
pyyaml==5.3.1             # via google-python-cloud-debugger
requests==2.24.0          # via google-api-core, google-cloud-profiler
rsa==4.6                  # via google-auth
scikit-learn==0.23.1      # via -r requirements.in
scipy==1.5.1              # via -r requirements.in, scikit-learn
six==1.15.0               # via google-api-core, google-api-python-client, google-auth, google-auth-httplib2, google-python-cloud-debugger, grpcio, protobuf, python-dateutil
threadpoolctl==2.1.0      # via scikit-learn
uritemplate==3.0.1        # via google-api-python-client
urllib3==1.25.9           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
